Make server modular
